requirements.txt
fastapi
uvicorn
langchain
pinecone-client
huggingface-hub
sentence-transformers
pypdf
requests
